+++ packages/kumo-relational-engine/src/kumo_relational_engine/pql/grammar/PQLGrammarVisitor.py
@@ -2,9 +2,9 @@
 # All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-# Generated from PQLGrammar.g4 by ANTLR 4.9.3
+# Generated from PQLGrammar.g4 by ANTLR 4.13.2
 from antlr4 import *
-if __name__ is not None and "." in __name__:
+if "." in __name__:
     from .PQLGrammarParser import PQLGrammarParser
 else:
     from PQLGrammarParser import PQLGrammarParser